Move attende api calls to attendee.ts
